Factor out dependency graph
